requirements.txt
bokeh==3.3.4
docarray==0.40.0
hvplot==0.9.2
langchain==0.1.7
langchain_community==0.0.20
langchain_core==0.1.23
langchain_openai==0.0.6
openai==1.12.0
panel==1.3.8
param==2.0.2
pydantic==2.6.1
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.31.1
wikipedia==1.4.0
